+++ aws-android-sdk-cognitoidentityprovider/src/main/java/com/amazonaws/mobileconnectors/cognitoidentityprovider/util/CognitoServiceConstants.java
@@ -29,6 +29,7 @@ public class CognitoServiceConstants {
     public static String CHLG_TYPE_CUSTOM_CHALLENGE                = "CUSTOM_CHALLENGE";
     public static String CHLG_TYPE_DEVICE_SRP_AUTH                 = "DEVICE_SRP_AUTH";
     public static String CHLG_TYPE_DEVICE_PASSWORD_VERIFIER        = "DEVICE_PASSWORD_VERIFIER";
+    public static String CHLG_TYPE_NEW_PASSWORD_REQUIRED           = "NEW_PASSWORD_REQUIRED";
     public static String AUTH_PARAM_REFRESH_TOKEN                  = "REFRESH_TOKEN";
     public static String AUTH_PARAM_DEVICE_KEY                     = "DEVICE_KEY";
     public static String AUTH_PARAM_SECRET_HASH                    = "SECRET_HASH";
@@ -45,6 +46,9 @@ public class CognitoServiceConstants {
     public static String CHLG_PARAM_CODE_DEL_MEDIUM                = "CODE_DELIVERY_DELIVERY_MEDIUM";
     public static String CHLG_PARAM_CODE_DEL_DESTINATION           = "CODE_DELIVERY_DESTINATION";
     public static String CHLG_PARAM_USER_ID_FOR_SRP                = "USER_ID_FOR_SRP";
+    public static String CHLG_PARAM_USER_ATTRIBUTE                 = "userAttributes";
+    public static String CHLG_PARAM_USER_ATTRIBUTE_PREFIX          = "userAttributes.";
+    public static String CHLG_PARAM_REQUIRED_ATTRIBUTE             = "requiredAttributes";
     public static String CHLG_RESP_SMS_MFA_CODE                    = "SMS_MFA_CODE";
     public static String CHLG_RESP_TIMESTAMP                       = "TIMESTAMP";
     public static String CHLG_RESP_PASSWORD_CLAIM_SECRET_BLOCK     = "PASSWORD_CLAIM_SECRET_BLOCK";
@@ -55,4 +59,5 @@ public class CognitoServiceConstants {
     public static String CHLG_RESP_ANSWER                          = "ANSWER";
     public static String CHLG_RESP_DEVICE_KEY                      = "DEVICE_KEY";
     public static String CHLG_RESP_PASSWORD                        = "PASSWORD";
+    public static String CHLG_RESP_NEW_PASSWORD                    = "NEW_PASSWORD";
 }